requirements.txt
flask==3.1.3
flask-cors==6.0.2
pandas==2.3.3
numpy==2.4.2
scikit-learn==1.8.0
matplotlib==3.10.8
joblib==1.5.3
PyJWT==2.12.0
pymongo==4.16.0
werkzeug==3.1.6
PyPDF2==3.0.1
python-docx==1.1.2
